added text option back in
