+++ BadgeLedApp.xcodeproj/project.xcworkspace/xcuserdata/gustav.xcuserdatad/Bookmarks/bookmarks.plist
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>top-level-items</key>
+	<array/>
+</dict>
+</plist>
